Script: 停止 Jar 程序时，判断程序状态
当前因为未判断值是否存在，导致如果程序未运行，运行此脚本停止程序时会报错

+++ Script/service.sh
@@ -78,7 +78,9 @@ function start() {
 }
 
 function stop() {
-    kill ${PROCESSID}
+    if [ -n ${PROCESSID} ]; then
+        kill ${PROCESSID}
+    fi
 }
 
 function restart() {
